+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - Add support for PS256, PS384 and PS512
 - Change API for both sign/verify to take a `Key` enum rather than bytes
+- Update to 2018 edition
 
 ## 6.0.1 (2019-05-10)